Rename the checksum verification file and package path
The root file and the package that held the verification logic were both
named checksum, which made it hard to tell which one a reference meant.

The root file becomes integrity.go: it holds the mode setting and the
Dewy methods that drive verification, not the checksum format itself.

internal/checksum moves to checksum, next to artifact, cache, registry
and the other packages of the same kind. It is used by both the registry
package and the deploy path, so it is not an implementation detail of
either one.
