+++ checksum/checksum.go
